requirements.txt
apscheduler==3.9.1.post1
Flask==2.2.2
flask_cors==3.0.10
python-dotenv==0.21.0
requests==2.25.1
